+++ src/main/java/seedu/address/logic/commands/MemberCommand/AddMemberCommand.java
@@ -6,54 +6,48 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
-import static seedu.address.logic.parser.MemberParser.CliSyntaxmember.PREFIX_MAJOR;
-import static seedu.address.logic.parser.MemberParser.CliSyntaxmember.PREFIX_POSTALCODE;
 
 import seedu.address.logic.CommandHistory;
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.members.member;
 import seedu.address.model.person.Person;
 
 /**
  * Adds a person to the address book.
  */
+public class AddCommand extends Command {
 
-public class AddMemberCommand{
-
-    public static final String COMMAND_WORD = "addmember";
+    public static final String COMMAND_WORD = "add";
     public static final String COMMAND_ALIAS = "a";
 
-    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Adds a member to the address book. "
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Adds a person to the address book. "
             + "Parameters: "
             + PREFIX_NAME + "NAME "
             + PREFIX_PHONE + "PHONE "
             + PREFIX_EMAIL + "EMAIL "
             + PREFIX_ADDRESS + "ADDRESS "
-            + PREFIX_POSTALCODE + "POSTALCODE "
-            + PREFIX_MAJOR + "MAJOR"
             + "[" + PREFIX_TAG + "TAG]...\n"
             + "Example: " + COMMAND_WORD + " "
             + PREFIX_NAME + "John Doe "
             + PREFIX_PHONE + "98765432 "
             + PREFIX_EMAIL + "johnd@example.com "
             + PREFIX_ADDRESS + "311, Clementi Ave 2, #02-25 "
-            + PREFIX_POSTALCODE + "609653 "
-            + PREFIX_MAJOR + "Computer engineering "
             + PREFIX_TAG + "friends "
             + PREFIX_TAG + "owesMoney";
 
-    public static final String MESSAGE_SUCCESS = "New member added: %1$s";
-    public static final String MESSAGE_DUPLICATE_PERSON = "This member already exists in the address book";
+    public static final String MESSAGE_SUCCESS = "New person added: %1$s";
+    public static final String MESSAGE_DUPLICATE_PERSON = "This person already exists in the address book";
 
-    private final member toAdd;
+    private final Person toAdd;
+
     /**
-     * Creates an AddCommand to add the specified {@code Member}
+     * Creates an AddCommand to add the specified {@code Person}
      */
-    public AddCommand(member member ) {
-
-        requireNonNull(member);
-        toAdd = member;
+    public AddCommand(Person person) {
+        requireNonNull(person);
+        toAdd = person;
     }
 
     @Override
@@ -76,7 +70,3 @@ public class AddMemberCommand{
                 && toAdd.equals(((AddCommand) other).toAdd));
     }
 }
-
-
-}
-
