--- src/main/java/seedu/address/logic/parser/MemberParser/CliSyntaxmember.java
@@ -1,17 +1,0 @@
-package seedu.address.logic.parser.MemberParser;
-
-import seedu.address.logic.parser.Prefix;
-
-public class CliSyntaxmember {
-
-    public static final Prefix PREFIX_NAME = new Prefix("n/");
-    public static final Prefix PREFIX_PHONE = new Prefix("p/");
-    public static final Prefix PREFIX_EMAIL = new Prefix("e/");
-    public static final Prefix PREFIX_ADDRESS = new Prefix("a/");
-    public static final Prefix PREFIX_TAG = new Prefix("t/");
-    public static final Prefix PREFIX_DATE = new Prefix("d/");
-    public static final Prefix PREFIX_POSTALCODE = new Prefix("c/");
-    public static final Prefix PREFIX_MAJOR = new Prefix ("m/");
-
-}
-
