+++ src/main/java/seedu/address/model/member/Email.java
