+++ src/main/java/seedu/address/model/member/Major.java
