--- src/main/java/seedu/address/model/members/Address.java
@@ -1,4 +1,0 @@
-package seedu.address.model.members;
-
-public class Address {
-}
