--- src/main/java/seedu/address/model/members/Name.java
@@ -1,4 +1,0 @@
-package seedu.address.model.members;
-
-public class Name {
-}
