+++ src/test/java/seedu/address/model/member/EmailTest.java
