+++ src/test/java/systemtests/FindMemberCommandSystemTest.java
@@ -14,8 +14,8 @@ import java.util.List;
 import org.junit.Test;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.FindCommand;
+import seedu.address.logic.commands.MemberCommand.DeleteCommand;
+import seedu.address.logic.commands.MemberCommand.FindCommand;
 import seedu.address.logic.commands.RedoCommand;
 import seedu.address.logic.commands.UndoCommand;
 import seedu.address.model.Model;
